Extract methods into SmartAnswersControllerTestHelper
I want to split up `SmartAnswersControllerTest` and this will make that easier.
